Refactor `Size` to provide a const interface
Currently, the support for const functions in Rust is very
limited, and does not support `match` nor `if`. This rules out
the possibility to have a const implementation of `Size`, if
`Size` is an enum.

Refactors `Size` into a struct that behaves like an enum, using
the value 0 as a marker for variable-sized bytewise types.